requirements.txt
absl-py==0.14.1
alabaster==0.7.12
argon2-cffi==21.1.0
astor==0.8.1
async-generator==1.10
attrs==21.2.0
Babel==2.9.1
backcall==0.2.0
backports.weakref==1.0rc1
bleach==1.5.0
certifi==2017.7.27.1
cffi==1.14.6
charset-normalizer==2.0.6
click==6.7
cycler==0.10.0
Cython==0.29.24
dataclasses==0.8
decorator==4.1.2
defusedxml==0.7.1
distro==1.6.0
docutils==0.17.1
dominate==2.6.0
entrypoints==0.2.3
Flask==0.12.2
Flask-Bootstrap==3.3.7.1
gast==0.5.2
grpcio==1.41.0
gunicorn==20.0.4
h5py==2.10.0
html5lib==0.9999999
idna==3.2
imageio==2.9.0
imagesize==1.2.0
imgaug==0.4.0
importlib-metadata==4.8.1
ipykernel==4.6.1
ipyparallel==7.1.0
ipython==6.1.0
ipython-genutils==0.2.0
ipywidgets==7.0.0
itsdangerous==0.24
jedi==0.10.2
Jinja2==2.9.6
jsonschema==2.6.0
jupyter-client==7.0.6
jupyter-core==4.8.1
jupyterlab-pygments==0.1.2
jupyterlab-widgets==1.0.2
Keras==2.0.8
Keras-Applications==1.0.8
Keras-Preprocessing==1.1.2
kiwisolver==1.3.1
Markdown==2.2.0
MarkupSafe==1.1.1
matplotlib==2.0.2
mistune==0.7.4
mock==4.0.3
nbclient==0.5.4
nbconvert==5.2.1
nbformat==4.4.0
nest-asyncio==1.5.1
networkx==2.5.1
nose==1.3.7
notebook==5.2.2
numpy==1.19.5
olefile==0.44
opencv-python==4.5.3.56
packaging==21.0
pandas==1.1.5
pandocfilters==1.4.2
parso==0.8.2
pexpect==4.2.1
pickleshare==0.7.4
Pillow==8.3.2
prometheus-client==0.11.0
prompt-toolkit==1.0.15
protobuf==3.18.1
psutil==5.8.0
ptyprocess==0.7.0
pycparser==2.20
Pygments==2.10.0
pyparsing==2.4.7
pyrsistent==0.18.0
python-dateutil==2.6.1
pytz==2017.2
PyWavelets==1.1.1
PyYAML==5.4.1
pyzmq==22.3.0
qtconsole==4.3.1
QtPy==1.11.2
requests==2.26.0
scikit-build==0.12.0
scikit-image==0.17.2
scipy==1.5.4
Send2Trash==1.8.0
Shapely==1.7.1
simplegeneric==0.8.1
six==1.16.0
snowballstemmer==2.1.0
Sphinx==4.2.0
sphinxcontrib-applehelp==1.0.2
sphinxcontrib-devhelp==1.0.2
sphinxcontrib-htmlhelp==2.0.0
sphinxcontrib-jsmath==1.0.1
sphinxcontrib-qthelp==1.0.3
sphinxcontrib-serializinghtml==1.1.5
tensorboard==1.13.1
tensorflow==1.13.1
tensorflow-estimator==1.13.0
termcolor==1.1.0
terminado==0.6
testpath==0.3.1
tifffile==2020.9.3
tornado==4.5.1
tqdm==4.62.3
traitlets==4.3.2
typing-extensions==3.10.0.2
urllib3==1.22
visitor==0.1.3
wcwidth==0.1.7
webencodings==0.5.1
Werkzeug==0.15.3
widgetsnbextension==3.0.8
zipp==3.6.0
